ServerApp/app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
aniso8601==9.0.1
apispec==5.1.1
apispec-webframeworks==0.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dateparser==1.1.1
flasgger==0.9.5
Flask==2.0.3
flask-apispec==0.11.0
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mistune==2.0.2
multidict==6.0.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
tzdata==2022.1
tzlocal==4.2
ujson==5.1.0
urllib3==1.26.9
uWSGI==2.0.20
webargs==8.1.0
websockets==9.1
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
Flask-Cors==3.0.10
